+++ server/src-lib/Hasura/App.hs
@@ -14,6 +14,7 @@ import           Options.Applicative
 import           System.Environment                   (getEnvironment, lookupEnv)
 import           System.Exit                          (exitFailure)
 
+import qualified Control.Concurrent.Async             as Async
 import qualified Control.Concurrent.Async.Lifted.Safe as LA
 import qualified Control.Concurrent.Extended          as C
 import qualified Data.Aeson                           as A
@@ -26,7 +27,6 @@ import qualified Database.PG.Query                    as Q
 import qualified Network.HTTP.Client                  as HTTP
 import qualified Network.HTTP.Client.TLS              as HTTP
 import qualified Network.Wai.Handler.Warp             as Warp
-import qualified System.Posix.Signals                 as Signals
 import qualified Text.Mustache.Compile                as M
 
 import           Hasura.Db
@@ -106,11 +106,12 @@ mkPGLogger (Logger logger) (Q.PLERetryMsg msg) =
 -- | most of the required types for initializing graphql-engine
 data InitCtx
   = InitCtx
-  { _icHttpManager :: !HTTP.Manager
-  , _icInstanceId  :: !InstanceId
-  , _icLoggers     :: !Loggers
-  , _icConnInfo    :: !Q.ConnInfo
-  , _icPgPool      :: !Q.PGPool
+  { _icHttpManager   :: !HTTP.Manager
+  , _icInstanceId    :: !InstanceId
+  , _icLoggers       :: !Loggers
+  , _icConnInfo      :: !Q.ConnInfo
+  , _icPgPool        :: !Q.PGPool
+  , _icShutdownLatch :: !ShutdownLatch
   }
 
 -- | Collection of the LoggerCtx, the regular Logger and the PGLogger
@@ -141,6 +142,7 @@ initialiseCtx hgeCmd rci = do
   httpManager <- liftIO $ HTTP.newManager HTTP.tlsManagerSettings
   instanceId <- liftIO generateInstanceId
   connInfo <- liftIO procConnInfo
+  latch <- liftIO newShutdownLatch
   (loggers, pool) <- case hgeCmd of
     -- for server command generate a proper pool
     HCServe so@ServeOptions{..} -> do
@@ -158,7 +160,7 @@ initialiseCtx hgeCmd rci = do
       pool <- getMinimalPool pgLogger connInfo
       pure (l, pool)
 
-  pure (InitCtx httpManager instanceId loggers connInfo pool, initTime)
+  pure (InitCtx httpManager instanceId loggers connInfo pool latch, initTime)
   where
     procConnInfo =
       either (printErrExit . connInfoErrModifier) return $ mkConnInfo rci
@@ -178,6 +180,21 @@ runTxIO :: Q.PGPool -> Q.TxMode -> Q.TxE QErr a -> IO a
 runTxIO pool isoLevel tx = do
   eVal <- liftIO $ runExceptT $ Q.runTx pool isoLevel tx
   either printErrJExit return eVal
+
+-- | A latch for the graceful shutdown of a server process.
+newtype ShutdownLatch = ShutdownLatch { unShutdownLatch :: C.MVar () }
+
+newShutdownLatch :: IO ShutdownLatch
+newShutdownLatch = fmap ShutdownLatch C.newEmptyMVar
+
+-- | Block the current thread, waiting on the latch.
+waitForShutdown :: ShutdownLatch -> IO ()
+waitForShutdown = C.takeMVar . unShutdownLatch
+
+-- | Initiate a graceful shutdown of the server associated with the provided 
+-- latch.
+shutdownGracefully :: InitCtx -> IO ()
+shutdownGracefully = flip C.putMVar () . unShutdownLatch . _icShutdownLatch
 
 runHGEServer
   :: ( HasVersion
@@ -295,23 +312,17 @@ runHGEServer ServeOptions{..} InitCtx{..} initTime = do
     runUnlockTx pool tx =
       liftIO $ runExceptT $ Q.runTx pool (Q.ReadCommitted, Nothing) tx
 
-    -- | Catches the SIGTERM signal and initiates a graceful shutdown. Graceful shutdown for regular HTTP
-    -- requests is already implemented in Warp, and is triggered by invoking the 'closeSocket' callback.
-    -- We only catch the SIGTERM signal once, that is, if the user hits CTRL-C once again, we terminate
-    -- the process immediately.
+    -- | Waits for the shutdown latch 'MVar' to be filled, and then
+    -- shuts down the server and associated resources.
+    -- Structuring things this way lets us decide elsewhere exactly how
+    -- we want to control shutdown. 
     shutdownHandler :: Logger Hasura -> IO () -> IO () -> IO ()
     shutdownHandler (Logger logger) shutdownApp closeSocket =
-      void $ Signals.installHandler
-        Signals.sigTERM
-        (Signals.CatchOnce shutdownSequence)
-        Nothing
-     where
-      shutdownSequence = do
+      void . Async.async $ do 
+        waitForShutdown _icShutdownLatch
+        logger $ mkGenericStrLog LevelInfo "server" "gracefully shutting down server"
         closeSocket
         shutdownApp
-        logShutdown
-
-      logShutdown = logger $ mkGenericStrLog LevelInfo "server" "gracefully shutting down server"
 
 runAsAdmin
   :: (MonadIO m)